Rename src to backend for future changes. Add shell scripts for easier dev starting
