+++ .gitignore
@@ -1,4 +1,5 @@
 certs/
 
+.idea*
 .env
 !.env-template
