+++ start.sh
@@ -1,0 +1,1 @@
+docker-compose -f docker-compose.dev.yml up && docker logs -f --since 5s msmf_dev
